oneoff form migration: fees section
